Une as funções de contagem de elementos em um único arquivo
