+++ src/cypress/e2e/opportunity/index.cy.js
@@ -1,6 +1,5 @@
 const { clearAllFilters } = require("../../commands/clearAllFilters");
-const { checkOpportunityCount } = require("../../commands/checkOpportunityCount");
-const { checkOpportunityCountWithClear } = require("../../commands/checkOpportunityCountWithClear");
+const { checkFilterCountOf } = require("../../commands/checkFilterCountOf");
 
 describe("Opportunity Page", () => {
     beforeEach(() => {
@@ -41,7 +40,7 @@ describe("Opportunity Page", () => {
 
         cy.wait(1000);
 
-        checkOpportunityCount();
+        checkFilterCountOf("opportunity", false);
     });
 
     it("Garante que os filtros por status das oportunidades funcionam", () => {
@@ -55,13 +54,13 @@ describe("Opportunity Page", () => {
 
         cy.wait(1000);
 
-        checkOpportunityCount();
+        checkFilterCountOf("opportunity", false);
 
         cy.get('.form > :nth-child(1) > :nth-child(4)').click();
 
         cy.wait(1000);
 
-        checkOpportunityCount();
+        checkFilterCountOf("opportunity", false);
     });
 
     it("Garante que o filtro de oportunidades de editais oficiais funciona", () => {
@@ -74,7 +73,7 @@ describe("Opportunity Page", () => {
         cy.get(".verified > input").click();
         cy.wait(1000);
 
-        checkOpportunityCount();
+        checkFilterCountOf("opportunity", false);
     });
 
     it("Garante que os filtros por tipo de oportunidade funcionam", () => {
@@ -89,7 +88,7 @@ describe("Opportunity Page", () => {
 
         cy.wait(1000);
 
-        checkOpportunityCount();
+        checkFilterCountOf("opportunity", false);
 
         cy.reload();
 
@@ -103,7 +102,7 @@ describe("Opportunity Page", () => {
 
         cy.wait(1000);
 
-        checkOpportunityCount();
+        checkFilterCountOf("opportunity", false);
     });
 
     it("Garante que os filtros por área de interesse funcionam", () => {
@@ -118,7 +117,7 @@ describe("Opportunity Page", () => {
 
         cy.wait(1000);
 
-        checkOpportunityCount();
+        checkFilterCountOf("opportunity", false);
     });
 
     it("Garante que o botão limpar filtros na pagina de oportunidades funciona", () => {
@@ -137,6 +136,6 @@ describe("Opportunity Page", () => {
 
         cy.wait(1000);
 
-        checkOpportunityCountWithClear();
+        checkFilterCountOf("opportunity", true);
     });
 });
