+++ src/cypress/e2e/spacesPage/compact.cy.js
@@ -18,7 +18,7 @@ describe("Pagina de Espaços", () => {
     });
 
     it("clica em \"Acessar\" e entra na pagina no espaço selecionado", () => {
-        cy.get('[href="https://mapas.tec.br/espaco/11/"]').last().click();
+        cy.get(`[href="${Cypress.config().baseUrl}/espaco/11/"]`).last().click();
         cy.url().should("include", "/espaco/");
         cy.contains('h1', 'Teatro Dulcina de Moraes');
     });
